pyproject.toml
[tool.poetry]
name = "profitgate-load"
version = "0.1.0"
description = ""
authors = ["yuraisme <yuraisme@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
loguru = "^0.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
